+++ bindings/nostr-sdk-ffi/src/protocol/nips/nip10.rs
@@ -14,8 +14,6 @@ pub enum Marker {
     Reply,
     /// Mention
     Mention,
-    /// Custom
-    Custom { custom: String },
 }
 
 impl From<Marker> for nip10::Marker {
@@ -24,7 +22,6 @@ impl From<Marker> for nip10::Marker {
             Marker::Root => Self::Root,
             Marker::Reply => Self::Reply,
             Marker::Mention => Self::Mention,
-            Marker::Custom { custom } => Self::Custom(custom),
         }
     }
 }
@@ -35,7 +32,6 @@ impl From<nip10::Marker> for Marker {
             nip10::Marker::Root => Self::Root,
             nip10::Marker::Reply => Self::Reply,
             nip10::Marker::Mention => Self::Mention,
-            nip10::Marker::Custom(custom) => Self::Custom { custom },
         }
     }
 }