+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 
 ### Removed
 
+* nostr: remove `Marker::Custom` variant ([Yuki Kishimoto])
 * pool: remove `Relay::support_negentropy` ([Yuki Kishimoto])
 
 ## [v0.36.0]